[v2] Move IDTransforms into its own file
